+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 ### Features
-- add authorization feature - `@Authorized` decorator and `authChecker` function in schema options ([see docs](https://github.com/19majkel94/type-graphql/tree/master/authorization.md))
+- add authorization feature - `@Authorized` decorator and `authChecker` function in schema options ([see docs](https://github.com/19majkel94/type-graphql/blob/master/docs/authorization.md))
 
 ## v0.6.0
 ### Features